requirements.txt
vk
tqdm
pandas
datetime
plotly
pymongo
python-telegram-bot